+++ ChangeLog.md
@@ -10,6 +10,9 @@ Changes that affect Michelson code are tagged with `[CONTRACT]`.
 
 Unreleased
 ==========
+
+0.3.1
+=====
 * [#105](https://github.com/serokell/tezos-btc/pull/105)
   - Add global option "--multisig-addr" to override the default multisig address/alias.
 * [#104](https://github.com/serokell/tezos-btc/pull/104)